requirements.txt
Flask==0.12
Flask-Testing
Flask-SQLAlchemy
SQLAlchemy-Utils==0.32.14
SQLAlchemy==1.1.10
psycopg2==2.7.1
flask-marshmallow==0.7.0
marshmallow-sqlalchemy
itsdangerous

#tests
pytest-pep8==1.0.6
pytest-flakes==2.0.0
pytest-cov==2.5.1
pyflakes==1.5.0
pytest==3.1.3